Create concept of a V3 entity
